adding design recipe

+++ planets.py
@@ -1,5 +1,15 @@
+#
+# Name: Jonathan Shan
+# Student ID:
+# Date (last modified): 9/21/18
+#
+# Lab 0 
+# Section 15
+# Purpose of Lab: Convert weight on different planets
+# 
+
 def weight_on_planets():
-	weight = float(input("What do you weigh on earth? "))
+	weight = int(input("What do you weigh on earth? "))
 	w_mars = weight * .38
 	w_jupiter = weight * 2.34
 	print("\n" + "On Mars you would weigh" , w_mars, "pounds." + "\n" + "On Jupiter you would weigh" , w_jupiter, "pounds.")
